Time: 4 ms (47.38%), Space: 89.4 MB (62.32%) - LeetHub

+++ 2358-number-of-ways-to-split-array/2358-number-of-ways-to-split-array.cpp
@@ -1,8 +1,8 @@
 class Solution {
 public:
     int waysToSplitArray(vector<int>& nums) {
-        long long l=0;
-        long long r=0;
+        long l=0;
+        long r=0;
         for(int i:nums){
             r+=i;
         }
